# docs/openapi.yaml
openapi: 3.1.0
info:
  title: oh-my-commic API
  version: "1.0.0"
  description: >
    面向亲子的 AI 漫画书应用后端 API。所有业务端点位于版本化前缀 /api/v1 下；
    健康探针 GET /api/health 保持不版本化（运维/CI 依赖其稳定路径）。
    认证使用 cookie session（登录后自动 set `session` cookie），前端请求需带
    credentials: 'include'。跨用户/不存在的资源一律返回 404（不泄露存在性）；
    AI/上游错误返回 502（504 超时 / 429 限流）；积分不足返回 402。
    错误响应统一为 `{"error": string}` 信封。图片经 /media/... 直接访问（不在本契约内）。
servers:
  - url: /
...
paths:
...
  /api/v1/books/{bookId}/upload:
    parameters:
      - $ref: "#/components/parameters/BookId"
    post:
      tags: [assets]
      summary: 上传资产图片（png/jpg/webp，≤5MB）
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: 上传成功，返回本地媒体 URL
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [imageUrl]
                properties:
                  imageUrl:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/books/{bookId}/characters:
    parameters:
      - $ref: "#/components/parameters/BookId"
    get:
      tags: [assets]
      summary: 列出书中的角色/宠物
      responses:
        "200":
          description: 角色列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Character"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [assets]
      summary: 新建角色/宠物
      description: >
        当 imageUrl 指向本书的本地上传（/media/{bookId}/...）时，后端会先漫画化
        （消耗积分），可能返回 402/429/502/504。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharacterInput"
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /api/v1/books/{bookId}/characters/{id}:
    parameters:
      - $ref: "#/components/parameters/BookId"
      - $ref: "#/components/parameters/Id"
    put:
      tags: [assets]
      summary: 更新角色/宠物
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharacterInput"
      responses:
        "200":
          description: 更新后的角色
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
    delete:
      tags: [assets]
      summary: 删除角色/宠物
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/books/{bookId}/scenes:
    parameters:
      - $ref: "#/components/parameters/BookId"
    get:
      tags: [assets]
      summary: 列出书中的场景
      responses:
        "200":
          description: 场景列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scene"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [assets]
      summary: 新建场景
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SceneInput"
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scene"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /api/v1/books/{bookId}/scenes/{id}:
    parameters:
      - $ref: "#/components/parameters/BookId"
      - $ref: "#/components/parameters/Id"
    put:
      tags: [assets]
      summary: 更新场景
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SceneInput"
      responses:
        "200":
          description: 更新后的场景
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scene"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
    delete:
      tags: [assets]
      summary: 删除场景
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    BookId:
      name: bookId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
  responses:
    OK:
      description: 操作成功
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [ok]
            properties:
              ok:
                type: boolean
    BadRequest:
      description: 请求参数/格式错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: 未登录
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    PaymentRequired:
      description: 积分不足
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 资源不存在或跨用户访问（不泄露存在性）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: 上游限流
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadGateway:
      description: AI/上游服务错误（不泄露上游细节）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    GatewayTimeout:
      description: 上游超时
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
...
    CharacterInput:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
          enum: [character, pet]
        name:
          type: string
        gender:
          type: string
        age:
          type: string
        personality:
          type: string
        description:
          type: string
        imageUrl:
          type: string
    Character:
      type: object
      additionalProperties: false
      required:
        - id
        - bookId
        - type
        - name
        - gender
        - age
        - personality
        - description
        - imageUrl
      properties:
        id:
          type: integer
          format: int64
        bookId:
          type: integer
          format: int64
        type:
          type: string
          enum: [character, pet]
        name:
          type: string
        gender:
          type: string
        age:
          type: string
        personality:
          type: string
        description:
          type: string
        imageUrl:
          type: string
    SceneInput:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
    Scene:
      type: object
      additionalProperties: false
      required: [id, bookId, name, description, imageUrl]
      properties:
        id:
          type: integer
          format: int64
        bookId:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
...
security:
  - sessionCookie: []